Conflicts:
	HiberLibros/src/main/java/com/hiberlibros/HiberLibros/controllers/InicioController.java
	HiberLibros/src/main/java/com/hiberlibros/HiberLibros/controllers/UsuarioController.java

+++ HiberLibros/src/main/java/com/hiberlibros/HiberLibros/services/UsuarioService.java
@@ -1,4 +1,3 @@
-
 package com.hiberlibros.HiberLibros.services;
 
 import com.hiberlibros.HiberLibros.entities.Usuario;
@@ -9,56 +8,59 @@ import org.springframework.stereotype.Service;
 import com.hiberlibros.HiberLibros.repositories.UsuarioRepository;
 import com.hiberlibros.HiberLibros.interfaces.UsuarioServiceI;
 import java.util.Optional;
+import javax.transaction.Transactional;
 
 /**
  *
  * @author Usuario
  */
 @Service
-public class UsuarioService implements UsuarioServiceI{
+public class UsuarioService implements UsuarioServiceI {
+
     @Autowired
     private UsuarioRepository urService;
-    
+
     @Autowired
     private ISeguridadService serviceUsuarioSeguridad;
 
-    
     @Override
     public String guardarUsuarioYSeguridad(Usuario u, String password) {
         String resultado = guardarUsuario(u);
-        Optional<Usuario> usu = urService.findByMail (u.getMail());
-        if (usu.isPresent()){
-            serviceUsuarioSeguridad.altaUsuarioSeguridad(  u.getMail(), u.getId(), password , "Usuario");
+        Optional<Usuario> usu = urService.findByMail(u.getMail());
+        if (usu.isPresent()) {
+            serviceUsuarioSeguridad.altaUsuarioSeguridad(u.getMail(), u.getId(), password, "Usuario");
         }
-        
+
         return resultado;
     }
-    
+
     @Override
     public String guardarUsuario(Usuario u) {
-        String resultado="";
-        int auxMail=u.getMail().indexOf("@");
-        String mailSubstring=u.getMail().substring(auxMail);//para comprar si el mail tiene buen formato
-        if(u.getNombre()==null || u.getApellido()==null || u.getDireccion()==null  || u.getCiudad()==null  || u.getMail()==null || u.getTelef()==null){
-            resultado="Error: Campo requerido vacío";
-        }
-        else if(!mailSubstring.contains(".")){ //@blabla. si no punto consideramos que no esta bien
-            resultado="Error: e-mail incorrecto";
-        }
-        else if(urService.findByMailContainsIgnoreCase(u.getMail()).isPresent()){//su ya existe ese mail
-            resultado="Error: Ya existe un usuario registrado con ese e-mail";
-        }
-        else{
+        String resultado = "";
+        int auxMail = u.getMail().indexOf("@");
+        String mailSubstring = u.getMail().substring(auxMail);//para comprar si el mail tiene buen formato
+        if (u.getNombre() == null || u.getApellido() == null || u.getDireccion() == null || u.getCiudad() == null || u.getMail() == null || u.getTelef() == null) {
+            resultado = "Error: Campo requerido vacío";
+        } else if (!mailSubstring.contains(".")) { //@blabla. si no punto consideramos que no esta bien
+            resultado = "Error: e-mail incorrecto";
+        } else if (urService.findByMailContainsIgnoreCase(u.getMail()).isPresent()) {//su ya existe ese mail
+            resultado = "Error: Ya existe un usuario registrado con ese e-mail";
+        } else {
             urService.save(u);
-            resultado="Usuario registrado con éxito";
+            resultado = "Usuario registrado con éxito";
         }
         return resultado;
 
     }
 
     @Override
+    @Transactional
     public void borrarUsuario(Integer id) {
-       urService.deleteById(id);
+        Optional<Usuario> usuario = urService.findById(id);
+        if (usuario.isPresent()) {
+            urService.deleteById(id);
+            serviceUsuarioSeguridad.bajaUsuarioSeguridadPorMail(usuario.get().getMail());
+        }
     }
 
     @Override
@@ -68,10 +70,9 @@ public class UsuarioService implements UsuarioServiceI{
 
     @Override
     public boolean registrado(String mail) { //comprueba si existe ese usuario por mail
-        if(urService.findByMail(mail).isEmpty()){
+        if (urService.findByMail(mail).isEmpty()) {
             return false;
-        }
-        else{
+        } else {
             return true;
         }
     }
@@ -92,5 +93,4 @@ public class UsuarioService implements UsuarioServiceI{
         return urService.findById(id).get();
     }
 
-    
 }
